+++ project-app/app/Http/Controllers/AsstController.php
@@ -473,54 +473,6 @@ class AsstController extends Controller
         return view($view, compact('retrieveData', 'fields', 'department', 'categories', 'location', 'model', 'status', 'manufacturer'));
     }
 
-
-
-
-
-    public function showRequestList()
-    {
-        // Fetch requests using the DB facade
-        $requests = DB::table('request')->where('request.requestor', '=', Auth::user()->id)
-            ->join('asset', 'asset.id', '=', 'request.asset_id')
-            ->join('department', 'department.id', '=', 'asset.dept_ID')
-            ->join('category', 'category.id', '=', 'asset.ctg_ID')
-            ->join('location', 'location.id', '=', 'asset.loc_key')
-            ->join('model', 'model.id', '=', 'asset.model_key')
-            ->select(
-                'asset.name',
-                'asset.id as asset_id', //remove nalang ni siya
-                'asset.image',
-                'asset.code',
-                'asset.cost',
-                'asset.depreciation',
-                'asset.salvageVal',
-                'asset.usage_Lifespan',
-                'asset.status',
-                'asset.custom_fields',
-                'asset.updated_at as assetCreated',
-                'asset.created_at as assetEdited',
-                'category.name as category',
-                'model.name as model',
-                'location.name as location',
-                'department.name as department',
-                'request.Description',
-                'request.id',
-                'request.status',
-                'request.requestor',
-                'request.approvedBy',
-                'request.created_at',
-                'request.updated_at',
-            )->get();
-
-        // Debugging the query output
-        if ($requests->isEmpty()) {
-            dd('No requests found in the database.');
-        }
-
-        // Pass the requests data to the view
-        return view('user.requestList', compact('requests'));
-    }
-
     public function downloadCsvTemplate()
     {
         // Define the readable column names that will replace foreign keys
