+++ project-app/resources/views/dept_head/asset.blade.php
@@ -186,12 +186,11 @@
                 })
                 .catch(error => console.log('Error:', error));
             });
-        }
 
         // DOMContentLoaded event to initialize all event listeners
         document.addEventListener('DOMContentLoaded', function () {
             const modalId = 'importModal';
-            
+
             // Modal open and close event listeners
             document.getElementById('openModalBtn').addEventListener('click', () => openModal(modalId));
             document.getElementById('closeModalBtn').addEventListener('click', () => closeModal(modalId));
